cudurru-services/requirements.txt
alembic==1.0.8
asn1crypto==0.24.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
backports.functools-lru-cache==1.5
bcrypt==3.1.7
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cheroot==6.5.4
CherryPy==18.1.1
cryptography==2.7
entrypoints==0.3
idna==2.8
isort==4.3.16
jaraco.functools==2.0
lazy-object-proxy==1.3.1
Mako==1.0.8
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.0.0
pluggy==0.9.0
portend==2.3
psycopg2==2.8.5
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.3.1
pylint-venv==1.1.0
pytest==6.1.0
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2018.9
PyYAML==5.1.1
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.1
tempora==1.14
typed-ast==1.4.0
urllib3==1.24.1
wrapt==1.11.1
zc.lockfile==1.4
GeoAlchemy2==0.8.4